wafr-agents/requirements.txt
# ============================================================================
# AgentCore CLI & Runtime (REQUIRED)
# ============================================================================
bedrock-agentcore-starter-toolkit
bedrock-agentcore

# ============================================================================
# Agent Framework (used by WAFR agents)
# ============================================================================
strands-agents
strands-agents-tools

# ============================================================================
# AWS SDK (for WA Tool integration and Bedrock API calls)
# ============================================================================
boto3>=1.34.0
botocore>=1.34.0

# ============================================================================
# Data Validation (used by agents for structured outputs)
# ============================================================================
pydantic>=2.0.0

# ============================================================================
# PDF Processing (optional - only if processing PDF transcripts)
# ============================================================================
# PyPDF2>=3.0.0
# pdfplumber>=0.10.0

# ============================================================================
# Security — JWT Authentication and Rate Limiting (Phase 3)
# ============================================================================
PyJWT[crypto]>=2.11.0
slowapi>=0.1.9

# ============================================================================
# Utilities
# ============================================================================
python-dateutil>=2.8.0
